Updated `OptionsContractPx.px_updated` to epoch sec

+++ rx_trade_ib_2/options/data/type/data.py
@@ -4,7 +4,7 @@ from rx_trade_ib_2.model.base import BasePydanticModel
 class OptionsContractPx(BasePydanticModel):
     ticker: str
     px: float | None = None
-    px_updated: int | None = None  # nanosecond
+    px_updated: float | None = None  # epoch sec
     open_interest: int
     iv: float
     delta: float | None = None
